move vault client to separate package, create amazon package
