+++ dao/amazon/secrets_manager_client.go
@@ -1,0 +1,1 @@
+package amazon
