+++ dao/interfaces.go
@@ -3,7 +3,6 @@ package dao
 import (
 	m "github.com/RedHatInsights/sources-api-go/model"
 	"github.com/RedHatInsights/sources-api-go/util"
-	"github.com/hashicorp/vault/api"
 	"github.com/redhatinsights/platform-go-middlewares/identity"
 )
 
@@ -155,13 +154,6 @@ type SourceTypeDao interface {
 	Update(src *m.SourceType) error
 	Delete(id *int64) error
 	GetByName(name string) (*m.SourceType, error)
-}
-
-type VaultClient interface {
-	Read(path string) (*api.Secret, error)
-	List(path string) (*api.Secret, error)
-	Write(path string, data map[string]interface{}) (*api.Secret, error)
-	Delete(path string) (*api.Secret, error)
 }
 
 type RhcConnectionDao interface {
